+++ pydidas/core/constants/generic_params_other.py
@@ -228,6 +228,15 @@ GENERIC_PARAMS_OTHER = {
             "number is always 0, irrespective of any offsets in the filenames."
         ),
     },
+    "detector_image_index": {
+        "type": int,
+        "default": 0,
+        "name": "Detector image number",
+        "choices": None,
+        "unit": "",
+        "allow_None": False,
+        "tooltip": ("The detector image number, as files are written to disk."),
+    },
     ################
     # Generic limits
     ################